// src/service/file/schemas/recentTaskResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { RecentTaskResultTaskType } from './recentTaskResultTaskType';
import type { RecentTaskResultSmartFolderApiType } from './recentTaskResultSmartFolderApiType';
import type { DriveItemResult } from './driveItemResult';
import type { LecturePlanResult } from './lecturePlanResult';
import type { LectureReportResult } from './lectureReportResult';
import type { StudentRecordResult } from './studentRecordResult';
import type { StoryBoardResult } from './storyBoardResult';

export interface RecentTaskResult {
  /** 드라이브 자료의 id */
  driveItemId: number;
  /** 드라이브 자료의 key */
  driveItemKey: string;
  /** 업무 종류. (LECTURE_PLAN:놀이계획, LECTURE_PLAN_REPORT:놀이보고서, STUDENT_RECORD:관찰기록, STORY_BOARD:스토리보드, STUDENT_CLASSIFICATION:아이분류, ACTIVITY_CLASSIFICATION: 활동분류, PHOTO_COMPOSITION: 아이합성, DATA_ENCRYPTION : 초상권보호SKETCH_CREATION : 도안 생성 PHOTO_ARRANGEMENT : 사진 정리,) */
  taskType: RecentTaskResultTaskType;
  /** 각 항목의 id. 업무 종류와 항목의 id가 한 쌍. */
  taskItemId: number;
  /** 소유자 계정 Id */
  ownerAccountId: number;
  /** 소유자 프로필 Id */
  ownerProfileId: number;
  /** 소유자 계정 Id */
  creatorAccountId: number;
  /** 소유자 프로필 Id */
  creatorProfileId: number;
  /** 생성일 */
  createdAt: string;
  /** 수정일 */
  modifiedAt: string;
  /**
   * 다른 썸네일이 있다면 url
   * @nullable
   */
  thumbUrl?: string | null;
  /**
   * 썸네일 표시용 사용자 프로필 조회용 코드
   * @nullable
   */
  userProfileCode?: string | null;
  /**
   * 썸네일 표시용 사용자 프로필 닉네임
   * @nullable
   */
  userProfileName?: string | null;
  /**
   * 썸네일 표시용 사용자 프로필 썸네일 url
   * @nullable
   */
  userProfileThumbUrl?: string | null;
  /**
   * 빠른 작업 ai라면 이동할 폴더의 스마트폴더 api 타입(아마 Photo고정일듯.)
   * @nullable
   */
  smartFolderApiType?: RecentTaskResultSmartFolderApiType;
  /**
   * 빠른 작업 ai라면 이동할 폴더의 스마트폴더의 id
   * @nullable
   */
  smartFolderId?: number | null;
  driveItem?: DriveItemResult;
  lecturePlan?: LecturePlanResult;
  lectureReport?: LectureReportResult;
  studentRecord?: StudentRecordResult;
  storyBoard?: StoryBoardResult;
}